+++ apps/ohsmart/package.json
@@ -14,6 +14,7 @@
     "@dans-framework/deposit": "workspace:*",
     "@dans-framework/layout": "workspace:*",
     "@dans-framework/pages": "workspace:*",
+    "@dans-framework/theme": "workspace:*",
     "@fontsource/roboto": "^5.0.7",
     "@mui/material": "^5.14.3",
     "i18next": "^23.4.1",
