+++ packages/theme/lib/index.tsx
@@ -1,0 +1,1 @@
+export { ThemeWrapper } from '../src/ThemeWrapper';
