+++ packages/theme/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@dans-framework/theme",
+  "main": "lib/index.tsx",
+  "version": "1.0.0",
+  "description": "",
+  "keywords": [],
+  "author": "",
+  "dependencies": {
+    "@mui/material": "^5.14.3",
+    "@mui/utils": "^5.14.7",
+    "react": "^18.2.0"
+  },
+  "devDependencies": {
+    "@types/react": "^18.2.15"
+  }
+}
